requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.0
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
chromedriver-binary==113.0.5672.24.0
chromedriver-binary-auto==0.2.5
Django==4.1.7
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
html5lib==1.1
idna==3.4
lxml==4.9.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.24.2
outcome==1.2.0
packaging==23.0
pandas==1.5.3
protobuf==3.20.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
selenium==4.8.2
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
undetected-chromedriver==3.4.6
urllib3==1.26.15
webdriver-manager==3.8.6
webencodings==0.5.1
websockets==10.4
wsproto==1.2.0
